modified onNextButtonClicked parameters type to int.

+++ Cupcake/app/src/main/java/com/example/cupcake/ui/StartOrderScreen.kt
@@ -49,7 +49,7 @@ import com.example.cupcake.data.DataSource
 @Composable
 fun StartOrderScreen(
     quantityOptions: List<Pair<Int, Int>>,
-    onNextButtonClicked: () -> Unit,
+    onNextButtonClicked: (Int) -> Unit,
     modifier: Modifier = Modifier
 ){
     Column(
